# docs/swagger.yaml
basePath: /
definitions:
...
  dto.GetGroupsResponse:
    properties:
      groups:
        items:
          $ref: '#/definitions/dto.GroupDTO'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationMeta'
    type: object
...
  dto.GroupDTO:
    properties:
      academic_group_id:
        type: integer
      academic_group_name:
        type: string
      admin_username:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  dto.ModeratorsResponse:
    properties:
      admin:
        $ref: '#/definitions/dto.UserDTO'
      moderators:
        items:
          $ref: '#/definitions/dto.UserDTO'
        type: array
    type: object
  dto.PaginationMeta:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      pages:
        example: 3
        type: integer
      total:
        example: 25
        type: integer
    type: object
...
  dto.UserDTO:
    properties:
      created_at:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  models.AcademicGroup:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.Group:
    properties:
      academic_group_id:
        description: Foreign key to AcademicGroup
        type: integer
      academicGroup:
        $ref: '#/definitions/models.AcademicGroup'
      admin:
        $ref: '#/definitions/models.User'
      adminID:
        description: Foreign key to User (admin)
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Subject:
    properties:
      academicGroup:
        $ref: '#/definitions/models.AcademicGroup'
      academicGroupID:
        type: integer
      createdAt:
        type: string
      name:
        type: string
      subjectID:
        type: integer
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      hashPassword:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
...
host: 185.221.155.133:8080
info:
  contact:
    email: vasyutin.i.a@edu.mirea.ru, saakovi.d.@edu.mirea.ru, pronin.p.v@edu.mirea.ru
    name: Иван Васютин, Павел Пронин, Давит Саакови
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: СВАГА
  version: "1.0"
paths:
...
  /api/groups:
    get:
      consumes:
      - application/json
      description: Retrieves a paginated list of groups with name, admin name, and
        academic group
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        in: query
        name: page_size
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetGroupsResponse'
        "400":
          description: Invalid page or page_size
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all groups with pagination
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Creates a group with the provided details, setting the authenticated
        user as admin
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Group data
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.Group'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.GroupDTO'
        "400":
          description: Invalid request body or missing required fields
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new group
      tags:
      - groups
  /api/groups/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a group by ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Group deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid group ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a group
      tags:
      - groups
    get:
      consumes:
      - application/json
      description: Retrieves a group with preloaded AcademicGroup and Admin data
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Invalid group ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a group by ID
      tags:
      - groups
    patch:
      consumes:
      - application/json
      description: Updates a group's name or academic_group_id, restricted to the
        group admin
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Group data (partial)
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.Group'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Invalid group ID, request body, or unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a group
      tags:
      - groups
  /api/groups/{id}/moderators:
    get:
      consumes:
      - application/json
      description: Get the admin and moderators of the specified group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ModeratorsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get group moderators and admin
      tags:
      - groups
  /api/groups/{id}/users:
    get:
      consumes:
      - application/json
      description: Get a list of users who are members of the specified group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get users in a group
      tags:
      - groups
...
  /api/groups/available:
    get:
      consumes:
      - application/json
      description: Get a paginated list of groups where the user is not a member,
        admin, or moderator
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetGroupsResponse'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get groups available to apply to
      tags:
      - groups
  /api/subjects:
    post:
      consumes:
      - application/json
      description: Creates a subject with the provided details
      parameters:
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Subject data
        in: body
        name: subject
        required: true
        schema:
          $ref: '#/definitions/models.Subject'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Subject'
        "400":
          description: Invalid request body or missing required fields
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new subject
      tags:
      - subjects
  /api/subjects/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a subject by ID
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Subject deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid subject ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Subject not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a subject
      tags:
      - subjects
    get:
      consumes:
      - application/json
      description: Retrieves a subject with preloaded Group data
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subject'
        "400":
          description: Invalid subject ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Subject not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a subject by ID
      tags:
      - subjects
    patch:
      consumes:
      - application/json
      description: Updates a subject's name, description, or group_id
      parameters:
      - description: Subject ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: Subject data (partial)
        in: body
        name: subject
        required: true
        schema:
          $ref: '#/definitions/models.Subject'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subject'
        "400":
          description: Invalid subject ID or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Subject not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a subject
      tags:
      - subjects
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"